backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary
django-cors-headers==4.4.0
django-environ